backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.3",
    "fastapi[standard]>=0.128.0",
    "passlib[argon2]>=1.7.4",
    "psycopg[binary]>=3.3.2",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.11.0",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "sqlmodel>=0.0.32",
]

[tool.basedpyright]
typeCheckingMode = "standard"

reportDeprecated = false
